requirements.txt
absl-py==2.3.1
astunparse==1.6.3
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
cloudpickle==3.1.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
descartes==1.1.0
dm-tree==0.1.8
fire==0.7.1
flatbuffers==25.9.23
fonttools==4.60.2
gast==0.7.0
google-auth==2.41.1
google-auth-oauthlib==1.2.3
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.14.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
joblib==1.5.2
keras==2.15.0
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
ml-dtypes==0.2.0
namex==0.1.0
numpy==1.26.4
nuscenes-devkit==1.2.0
nvidia-cublas-cu12==12.2.5.6
nvidia-cuda-cupti-cu12==12.2.142
nvidia-cuda-nvcc-cu12==12.2.140
nvidia-cuda-runtime-cu12==12.2.140
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.8.103
nvidia-curand-cu12==10.3.3.141
nvidia-cusolver-cu12==11.5.2.141
nvidia-cusparse-cu12==12.1.2.141
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
oauthlib==3.3.1
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
parameterized==0.9.0
pillow==11.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
Pygments==2.19.2
pyparsing==3.2.5
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
shapely==2.0.7
six==1.17.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-addons==0.23.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.23.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typeguard==2.13.3
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
wrapt==1.14.2
zipp==3.23.0
